Fix various dangling pointer issues.
The current code was assuming that memory pointed by current_bind_parameters
would get zeroed when required, but this isn't reliable and can be realiably
hit with a -DCLOBBER_FREED_MEMORY build.  Fix it by properly tracking the
executor depth and reset the variable when needed.

The current code was also assuming that the current memory context in
pel_log_error() wouldn't be reset in case of a subsequent error, but this also
can happen.  Fix by calling raw_parser() in a dedicated temporary context,
child of CurTransactionContext, that is reset right after the call.