Restart the main bgworker immediately.
Note that the main bgworker can hit an error and restart if the dynamic
bgworker fails, for instance if it tries to process entries for a database than
has been dropped since.  This scenario should be tested and fixed but for now
simply reduce the time needed for the main bgworker to restart in case of
problem.