Fix privilege issues on using pg_dbms_errlog as a basic user like Oracle DBMS_ERRLOG behavior.
A user must be granted the DML privileges to the table and to the error log
table to be able to use this feature. Insert to the registration table is done
internally by superuser. To allow a user to create an error logging table he
must be granted to execute the create_error_log() function and have read/write
access to the registration table dbms_errlog.register_errlog_tables.

Add regression test for privilege on errorlog table.